separate pedestian and crowd test cases into different modules
