+++ cabot_site_large_room/cabot_site_large_room/tests.py
@@ -1,7 +1,3 @@
-import random
-import time
-
-
 def config(tester):
     tester.config['init_x'] = 0.0
     tester.config['init_y'] = 0.0
@@ -77,6 +73,7 @@ def test_category1_case1_move_towards_a_pedestrian(tester):
             },
         },
     ])
+    _check_robot_on_person_collision(tester)
     _goto_target1(tester)
 
 
@@ -96,6 +93,7 @@ def test_category1_case2_move_towards_a_pedestrian_and_avoid(tester):
             },
         },
     ])
+    _check_robot_on_person_collision(tester)
     _goto_target1(tester)
 
 
@@ -122,6 +120,7 @@ def test_category1_case4_robot_overtaking(tester):
         topic_type="pedestrian_plugin_msgs/msg/Metric",
         condition="msg.name=='total_time' and 30.0<=msg.value"
     )
+    _check_robot_on_person_collision(tester)
     _goto_target1(tester)
 
 
@@ -141,6 +140,7 @@ def test_category1_case5_down_the_path(tester):
             },
         },
     ])
+    _check_robot_on_person_collision(tester)
     _goto_target1(tester)
 
 
@@ -181,513 +181,5 @@ def test_category3_case1_move_across_a_pedestrian_proceed(tester):
             },
         },
     ])
-    _goto_target1(tester)
-
-
-def test_category5_allocate_actors_preparation(tester):
-    nx = 10
-    ny = 5
-
-    init_x = 0.0
-    init_y = 10.0
-    for j in range(0, ny):
-        actors = []
-        for i in range(0, nx):
-            index = i + nx * j
-            print(f"index={index}")
-            actor = {
-                "name": f'actor{index}',
-                "module": "pedestrian.pool",
-                "params": {
-                    "init_x": init_x + i,
-                    "init_y": init_y + j,
-                    "init_a": -180.0,
-                    "velocity": 0.95,
-                },
-            }
-            actors.append(actor)
-        tester.setup_actors(actors=actors)
-        time.sleep(1)
-
-
-def test_category5_case2_move_parallel_traffic(tester):
-    # 5.2 parallel traffic
-    tester.check_collision()
-    tester.reset_position(x=-3.0)
-
-    actors = []
-
-    nx = 5
-    ny = 5
-
-    interval_x = 4.0
-    interval_y = 3.0
-
-    x_0 = 0.0
-    y_0 = 0.0
-    shift_y = -0.5
-
-    # forward flow
-    for i in range(0, nx):
-        for j in range(0, ny):
-            index = j + ny * i
-            init_x = x_0 - interval_x * nx + interval_x * i
-            init_y = y_0 - interval_y * (ny//2) + interval_y * j
-            actors.append({
-                "name": f'actor{index}',
-                "module": "pedestrian.walk_across",
-                "params": {
-                    "init_x": init_x,
-                    "init_y": init_y,
-                    "init_a": 0.0,
-                    "velocity": 1.0,
-                },
-            })
-
-    # inverse flow
-    for i in range(0, nx):
-        for j in range(0, ny):
-            index = j + ny * i + ny * nx
-            init_x = x_0 + interval_x * nx - interval_x * i
-            init_y = y_0 - interval_y * (ny//2) + interval_y * j + shift_y
-            actors.append({
-                "name": f'actor{index}',
-                "module": "pedestrian.walk_across",
-                "params": {
-                    "init_x": init_x,
-                    "init_y": init_y,
-                    "init_a": 180.0,
-                    "velocity": 1.0,
-                },
-            })
-
-    tester.setup_actors(actors=actors)
-    _goto_target1(tester)
-
-
-def test_category5_case2_move_parallel_traffic_repeat(tester):
-    # 5.2 parallel traffic (repeat)
-    tester.check_collision()
-    tester.reset_position(x=-6.0)
-
-    actors = []
-
-    nx = 5
-    ny = 5
-
-    interval_x = 4.0
-    interval_y = 3.0
-
-    x_0 = 0.0
-    y_0 = 0.0
-    shift_y = -0.5
-
-    goal_x_dist = 12
-
-    # forward flow
-    for i in range(0, nx):
-        for j in range(0, ny):
-            index = j + ny * i
-            init_x = x_0 - 8.0
-            init_y = y_0 - interval_y * (ny//2) + interval_y * j
-            goal_x = x_0 + goal_x_dist
-            goal_y = init_y
-            actors.append({
-                "name": f'actor{index}',
-                "module": "pedestrian.walk_across",
-                "params": {
-                    "init_x": init_x,
-                    "init_y": init_y,
-                    "init_a": 0.0,
-                    "velocity": 1.0,
-                    "goal_x": goal_x,
-                    "goal_y": goal_y,
-                    "repeat": 1,
-                    "start_after": i * interval_x
-                },
-            })
-
-    # inverse flow
-    for i in range(0, nx):
-        for j in range(0, ny):
-            index = j + ny * i + ny * nx
-            init_x = x_0 + 8.0
-            init_y = y_0 - interval_y * (ny//2) + interval_y * j + shift_y
-            goal_x = x_0 - goal_x_dist
-            goal_y = init_y
-            actors.append({
-                "name": f'actor{index}',
-                "module": "pedestrian.walk_across",
-                "params": {
-                    "init_x": init_x,
-                    "init_y": init_y,
-                    "init_a": 180.0,
-                    "velocity": 1.0,
-                    "goal_x": goal_x,
-                    "goal_y": goal_y,
-                    "repeat": 1,
-                    "start_after": i * interval_x
-                },
-            })
-
-    tester.setup_actors(actors=actors)
-    _goto_target1(tester)
-
-
-def test_category5_case3_move_perpendicular_traffic(tester):
-    # 5.3 Perpendicular traffic
-    tester.check_collision()
-    tester.reset_position(x=-10.0)
-
-    actors = []
-
-    interval_x = 3.0
-    interval_y = 3.0
-
-    init_x = -5.0
-    init_y = 0.0
-    for i in range(0, 5):
-        actor = {
-            "name": f'actor{i}',
-            "module": "pedestrian.walk_across",
-            "params": {
-                "init_x": init_x,
-                "init_y": init_y,
-                "init_a": -90.0,
-                "velocity": 0.95,
-            },
-        }
-        actors.append(actor)
-        init_y += interval_y
-
-    init_x = -5.0 + interval_x
-    init_y = -interval_x
-    for i in range(5, 10):
-        actor = {
-            "name": f'actor{i}',
-            "module": "pedestrian.walk_across",
-            "params": {
-                "init_x": init_x,
-                "init_y": init_y,
-                "init_a": 90.0,
-                "velocity": 0.95,
-            },
-        }
-        actors.append(actor)
-        init_y -= interval_y
-
-    init_x = -5.0 + interval_x + interval_x
-    init_y = interval_x + interval_x
-    for i in range(10, 15):
-        actor = {
-            "name": f'actor{i}',
-            "module": "pedestrian.walk_across",
-            "params": {
-                "init_x": init_x,
-                "init_y": init_y,
-                "init_a": -90.0,
-                "velocity": 0.95,
-            },
-        }
-        actors.append(actor)
-        init_y += interval_y
-
-    init_x = -5.0 + 3 * interval_x
-    init_y = - 3 * interval_x
-    for i in range(15, 20):
-        actor = {
-            "name": f'actor{i}',
-            "module": "pedestrian.walk_across",
-            "params": {
-                "init_x": init_x,
-                "init_y": init_y,
-                "init_a": 90.0,
-                "velocity": 0.95,
-            },
-        }
-        actors.append(actor)
-        init_y -= interval_y
-
-    tester.setup_actors(actors=actors)
-    _goto_target1(tester)
-
-
-def test_category5_case3_move_perpendicular_traffic_repeat(tester):
-    # 5.3 Perpendicular traffic (repeat)
-    tester.check_collision()
-    tester.reset_position(x=-10.0)
-
-    actors = []
-
-    interval_x = 3.0
-    interval_y = 3.0
-
-    init_x = -2.5
-    init_y = 7.5
-    goal_x = init_x
-    goal_y = -init_y
-    for i in range(0, 5):
-        actor = {
-            "name": f'actor{i}',
-            "module": "pedestrian.walk_across",
-            "params": {
-                "init_x": init_x,
-                "init_y": init_y,
-                "init_a": -90.0,
-                "velocity": 0.95,
-                "goal_x": goal_x,
-                "goal_y": goal_y,
-                "repeat": 1,
-                "start_after": i * interval_y
-            },
-        }
-        actors.append(actor)
-
-    init_x = -2.5 + interval_x
-    init_y = -7.5
-    goal_x = init_x
-    goal_y = -init_y
-    for i in range(0, 5):
-        actor = {
-            "name": f'actor{i+5}',
-            "module": "pedestrian.walk_across",
-            "params": {
-                "init_x": init_x,
-                "init_y": init_y,
-                "init_a": 90.0,
-                "velocity": 0.95,
-                "goal_x": goal_x,
-                "goal_y": goal_y,
-                "repeat": 1,
-                "start_after": i * interval_y
-            },
-        }
-        actors.append(actor)
-
-    init_x = -2.5 + 2 * interval_x
-    init_y = 7.5
-    goal_x = init_x
-    goal_y = -init_y
-    for i in range(0, 5):
-        actor = {
-            "name": f'actor{i+10}',
-            "module": "pedestrian.walk_across",
-            "params": {
-                "init_x": init_x,
-                "init_y": init_y,
-                "init_a": -90.0,
-                "velocity": 0.95,
-                "goal_x": goal_x,
-                "goal_y": goal_y,
-                "repeat": 1,
-                "start_after": i * interval_y
-            },
-        }
-        actors.append(actor)
-
-    init_x = -2.5 + 3 * interval_x
-    init_y = -7.5
-    goal_x = init_x
-    goal_y = -init_y
-    for i in range(0, 5):
-        actor = {
-            "name": f'actor{i+15}',
-            "module": "pedestrian.walk_across",
-            "params": {
-                "init_x": init_x,
-                "init_y": init_y,
-                "init_a": 90.0,
-                "velocity": 0.95,
-                "goal_x": goal_x,
-                "goal_y": goal_y,
-                "repeat": 1,
-                "start_after": i * interval_y
-            },
-        }
-        actors.append(actor)
-
-    tester.setup_actors(actors=actors)
-    _goto_target1(tester)
-
-
-def test_category6_case1_sfm_actors(tester):
-    # 6.1 crowd navigation
-    tester.check_collision()
-    bound = 5.0
-    actors = []
-    n_actors = 10
-    for i in range(0, n_actors):
-        actors.append({
-            "name": f"actor{i}",
-            "module": "pedestrian.walk_sfm",
-            "params": {
-                "init_x": random.uniform(-bound, bound),
-                "init_y": random.uniform(-bound, bound),
-                "init_a": random.uniform(-180.0, 180.0),
-                "velocity": 1.0,
-                "n_actors": n_actors,
-                "min_x": -bound,
-                "max_x": bound,
-                "min_y": -bound,
-                "max_y": bound,
-            },
-        })
-    tester.reset_position(x=-6.0)
-    tester.setup_actors(actors=actors)
-    _goto_target1(tester)
-
-
-def test_category6_case2_sfm_parallel_traffic(tester):
-    # 6.2 parallel traffic
-    tester.check_collision()
-    actors = []
-
-    nx = 5
-    ny = 5
-
-    interval_x = 3.0
-    interval_y = 2.0
-
-    x_0 = 0.0
-    y_0 = 0.0
-    shift_y = -0.5
-
-    n_actors = nx * ny * 2
-
-    std_x = 0.5
-    std_y = 0.5
-
-    # forward flow
-    for i in range(0, nx):
-        for j in range(0, ny):
-            index = j + ny * i
-            init_x = x_0 - interval_x * nx + interval_x * i + random.normalvariate(0.0, std_x)
-            init_y = y_0 - interval_y * (ny//2) + interval_y * j + random.normalvariate(0.0, std_y)
-            actors.append({
-                "name": f"actor{index}",
-                "module": "pedestrian.walk_sfm",
-                "params": {
-                    "init_x": init_x,
-                    "init_y": init_y,
-                    "init_a": 0.0,
-                    "velocity": 1.0,
-                    "n_actors": n_actors,
-                    "goal_x": 14.0,
-                    "goal_y": init_y
-                },
-            })
-
-    # inverse flow
-    for i in range(0, nx):
-        for j in range(0, ny):
-            index = j + ny * i + ny * nx
-            init_x = x_0 + interval_x * nx - interval_x * i + random.normalvariate(0.0, std_x)
-            init_y = y_0 - interval_y * (ny//2) + interval_y * j + shift_y + random.normalvariate(0.0, std_y)
-            actors.append({
-                "name": f"actor{index}",
-                "module": "pedestrian.walk_sfm",
-                "params": {
-                    "init_x": init_x,
-                    "init_y": init_y,
-                    "init_a": 180.0,
-                    "velocity": 1.0,
-                    "n_actors": n_actors,
-                    "goal_x": -14.0,
-                    "goal_y": init_y
-                },
-            })
-
-    tester.reset_position(x=-3.0)
-    tester.setup_actors(actors=actors)
-    _goto_target1(tester)
-
-
-def test_category6_case3_sfm_perpendicular_traffic(tester):
-    # 6.3 Perpendicular traffic
-    tester.check_collision()
-    tester.reset_position(x=-10.0)
-
-    actors = []
-
-    interval_x = 3.0
-    interval_y = 3.0
-    n_actors = 20
-
-    init_x = -5.0
-    init_y = 0.0
-    for i in range(0, 5):
-        actor = {
-            "name": f"actor{i}",
-            "module": "pedestrian.walk_sfm",
-            "params": {
-                "init_x": init_x,
-                "init_y": init_y,
-                "init_a": -90.0,
-                "velocity": 0.95,
-                "n_actors": n_actors,
-                "goal_x": init_x,
-                "goal_y": -15.0
-            }
-        }
-        actors.append(actor)
-        init_y += interval_y
-
-    init_x = -5.0 + interval_x
-    init_y = -interval_x
-    for i in range(5, 10):
-        actor = {
-            "name": f"actor{i}",
-            "module": "pedestrian.walk_sfm",
-            "params": {
-                "init_x": init_x,
-                "init_y": init_y,
-                "init_a": 90.0,
-                "velocity": 0.95,
-                "n_actors": n_actors,
-                "goal_x": init_x,
-                "goal_y": 15.0
-            }
-        }
-        actors.append(actor)
-        init_y -= interval_y
-
-    init_x = -5.0 + interval_x + interval_x
-    init_y = interval_x + interval_x
-    for i in range(10, 15):
-        actor = {
-            "name": f"actor{i}",
-            "module": "pedestrian.walk_sfm",
-            "params": {
-                "init_x": init_x,
-                "init_y": init_y,
-                "init_a": -90.0,
-                "velocity": 0.95,
-                "n_actors": n_actors,
-                "goal_x": init_x,
-                "goal_y": -15.0
-            }
-        }
-        actors.append(actor)
-        init_y += interval_y
-
-    init_x = -5.0 + 3 * interval_x
-    init_y = - 3 * interval_x
-    for i in range(15, 20):
-        actor = {
-            "name": f"actor{i}",
-            "module": "pedestrian.walk_sfm",
-            "params": {
-                "init_x": init_x,
-                "init_y": init_y,
-                "init_a": 90.0,
-                "velocity": 0.95,
-                "n_actors": n_actors,
-                "goal_x": init_x,
-                "goal_y": 15.0
-            }
-        }
-        actors.append(actor)
-        init_y -= interval_y
-
-    tester.setup_actors(actors=actors)
+    _check_robot_on_person_collision(tester)
     _goto_target1(tester)
